[2022.04.20] TEST 배포

+++ controller/store.js
@@ -6,22 +6,28 @@ module.exports = {
   storeBread: async (req, res, next) => {
     try {
       const { body, user_idx } = req;
-      // 제일 최근 데이터가 같은 데이터인지 체크
-      await client.connect();
 
-      const stroedBread = await client.get(`${user_idx}`);
+      // io 인지 체크
+      if (process.env.NODE_MODE == "IO") {
+        // 제일 최근 데이터가 같은 데이터인지 체크
+        await client.connect();
 
-      if (stroedBread == body.bread) {
-        console.log(stroedBread);
+        const stroedBread = await client.get(`${user_idx}`);
+
+        if (stroedBread == body.bread) {
+          console.log(stroedBread);
+        } else {
+          console.log(stroedBread);
+          body.user_idx = user_idx;
+          await db.store.create(body);
+          client.del(`${user_idx}`);
+          client.set(`${user_idx}`, `${body.bread}`);
+        }
+        await client.quit();
+        return res.send({ success: 200 });
       } else {
-        console.log(stroedBread);
-        body.user_idx = user_idx;
-        await db.store.create(body);
-        client.del(`${user_idx}`);
-        client.set(`${user_idx}`, `${body.bread}`);
+        return res.send({ success: 200 });
       }
-      await client.quit();
-      return res.send({ success: 200 });
     } catch (err) {
       next(err);
     }
